+++ sql/pg_duckdb--0.3.0--1.0.0.sql
@@ -587,6 +587,11 @@ CREATE CAST (duckdb.unresolved_type AS bytea)
 CREATE CAST (duckdb.unresolved_type AS bytea[])
     WITH INOUT;
 
+CREATE CAST (duckdb.unresolved_type AS text)
+    WITH INOUT;
+CREATE CAST (duckdb.unresolved_type AS text[])
+    WITH INOUT;
+
 CREATE OPERATOR pg_catalog.~ (
     LEFTARG = duckdb.unresolved_type,
     RIGHTARG = duckdb.unresolved_type,